refactor(public): Reubicación y corrección del componente de Avisos Escolares y vista del Calendario
